+++ src/inspect_ai/_cli/eval.py
@@ -6,8 +6,10 @@ from typing_extensions import Unpack
 
 from inspect_ai import Epochs, eval, eval_retry
 from inspect_ai._eval.evalset import eval_set
+from inspect_ai._util.config import resolve_args
 from inspect_ai._util.constants import (
     ALL_LOG_LEVELS,
+    DEFAULT_BATCH_SIZE,
     DEFAULT_EPOCHS,
     DEFAULT_LOG_LEVEL_TRANSCRIPT,
     DEFAULT_LOG_SHARED,
@@ -18,7 +20,7 @@ from inspect_ai._util.file import filesystem
 from inspect_ai._util.samples import parse_sample_id, parse_samples_limit
 from inspect_ai.log._file import log_file_info
 from inspect_ai.model import GenerateConfigArgs
-from inspect_ai.model._generate_config import ResponseSchema
+from inspect_ai.model._generate_config import BatchConfig, ResponseSchema
 from inspect_ai.scorer._reducer import create_reducers
 from inspect_ai.solver._solver import SolverSpec
 
@@ -28,6 +30,7 @@ from .common import (
     process_common_options,
 )
 from .util import (
+    int_bool_or_str_flag_callback,
     int_or_bool_flag_callback,
     parse_cli_args,
     parse_cli_config,
@@ -62,6 +65,7 @@ MAX_RETRIES_HELP = (
     "Maximum number of times to retry model API requests (defaults to unlimited)"
 )
 TIMEOUT_HELP = "Model API request timeout in seconds (defaults to no timeout)"
+BATCH_HELP = "Batch requests together to reduce API calls when using a model that supports batching (by default, no batching). Specify --batch to batch with default configuration,  specify a batch size e.g. `--batch=1000` to configure batches of 1000 requests, or pass the file path to a YAML or JSON config file with batch configuration."
 
 
 def eval_options(func: Callable[..., Any]) -> Callable[..., click.Context]:
@@ -468,6 +472,15 @@ def eval_options(func: Callable[..., Any]) -> Callable[..., click.Context]:
         envvar="INSPECT_EVAL_RESPONSE_SCHEMA",
     )
     @click.option(
+        "--batch",
+        is_flag=False,
+        flag_value="true",
+        default=None,
+        callback=int_bool_or_str_flag_callback(DEFAULT_BATCH_SIZE, None),
+        help=BATCH_HELP,
+        envvar="INSPECT_EVAL_BATCH",
+    )
+    @click.option(
         "--log-format",
         type=click.Choice(["eval", "json"], case_sensitive=False),
         envvar=["INSPECT_LOG_FORMAT", "INSPECT_EVAL_LOG_FORMAT"],
@@ -542,6 +555,7 @@ def eval_command(
     reasoning_summary: Literal["concise", "detailed", "auto"] | None,
     reasoning_history: Literal["none", "all", "last", "auto"] | None,
     response_schema: ResponseSchema | None,
+    batch: int | str | None,
     message_limit: int | None,
     token_limit: int | None,
     time_limit: int | None,
@@ -718,6 +732,7 @@ def eval_set_command(
     reasoning_summary: Literal["concise", "detailed", "auto"] | None,
     reasoning_history: Literal["none", "all", "last", "auto"] | None,
     response_schema: ResponseSchema | None,
+    batch: int | str | None,
     message_limit: int | None,
     token_limit: int | None,
     time_limit: int | None,
@@ -1000,6 +1015,9 @@ def config_from_locals(locals: dict[str, Any]) -> GenerateConfigArgs:
             if key == "response_schema":
                 if value is not None:
                     value = ResponseSchema.model_validate_json(value)
+            if key == "batch":
+                value = BatchConfig.model_validate(resolve_args(value))
+
             config[key] = value  # type: ignore
     return config