+++ src/inspect_ai/model/_model.py
@@ -32,7 +32,11 @@ from tenacity import (
 )
 from tenacity.stop import StopBaseT
 
-from inspect_ai._util.constants import DEFAULT_MAX_CONNECTIONS, HTTP
+from inspect_ai._util.constants import (
+    DEFAULT_MAX_CONNECTIONS,
+    DEFAULT_MAX_CONNECTIONS_BATCH,
+    HTTP,
+)
 from inspect_ai._util.content import (
     Content,
     ContentImage,
@@ -126,7 +130,6 @@ class ModelAPI(abc.ABC):
         """
         self.model_name = model_name
         self.base_url = base_url
-        self.config = config
 
         from inspect_ai.hooks._hooks import override_api_key
 
@@ -761,6 +764,8 @@ class Model:
         max_connections = (
             config.max_connections
             if config.max_connections
+            else DEFAULT_MAX_CONNECTIONS_BATCH
+            if config.batch
             else self.api.max_connections()
         )
         model_name = ModelName(self)